+++ lib/screens/adoptdetails.dart
