+++ lib/screens/adoptpage.dart
@@ -1,167 +1,294 @@
 import 'package:flutter/material.dart';
 
-class AdoptPage extends StatefulWidget {
-  const AdoptPage({super.key});
+
+class PetAdoptPage extends StatefulWidget {
+  const PetAdoptPage({super.key});
 
   @override
-  State<AdoptPage> createState() => _HomePageState();
+  State<PetAdoptPage> createState() => _PetAdoptPageState();
 }
 
-class _HomePageState extends State<AdoptPage> {
+class _PetAdoptPageState extends State<PetAdoptPage> {
+  final categories = ['Cat', 'Dog', 'Fish', 'Bird'];
+  String selectedCategory = 'Cat';
+
+  final adoptPets = const [
+    {'name': 'Ig', 'image': 'assets/images/cat3.jpeg', 'gender': 'Male'},
+    {'name': 'Lorna', 'image': 'assets/images/cat1.jpg', 'gender': 'Male'},
+    {'name': 'Max', 'image': 'assets/images/cat5.jpg', 'gender': 'Male'},
+    {'name': 'Bella', 'image': 'assets/images/cat2.jpg', 'gender': 'Female'},
+    {'name': 'Charlie', 'image': 'assets/images/cat4.jpg', 'gender': 'Male'},
+  ];
+
+  final suggestions = const [
+    {'name': 'Blessing', 'image': 'assets/images/dog3.jpg', 'gender': 'Female'},
+    {'name': 'Pumi', 'image': 'assets/images/fish2.jpg', 'gender': 'Female'},
+    {'name': 'Christian', 'image': 'assets/images/dog4.jpg', 'gender': 'Female'},
+  ];
+
+  final suggestions2 = const [
+    {'name': 'Toby', 'image': 'assets/images/dog2.jpg', 'gender': 'Male'},
+    {'name': 'Luna', 'image': 'assets/images/dog1.jpg', 'gender': 'Male'},
+    {'name': 'Oscar', 'image': 'assets/images/fish1.jpg', 'gender': 'Male'},
+  ];
+
+  List<bool> likedPets = List.generate(5, (index) => false);
+  List<bool> likedSuggestions = List.generate(3, (index) => false);
+  List<bool> likedSuggestions2 = List.generate(3, (index) => false);
+
+  void _toggleHeart(int index, String section) {
+    setState(() {
+      if (section == 'adopt') {
+        likedPets[index] = !likedPets[index];
+      } else if (section == 'suggestion1') {
+        likedSuggestions[index] = !likedSuggestions[index];
+      } else if (section == 'suggestion2') {
+        likedSuggestions2[index] = !likedSuggestions2[index];
+      }
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
+    return Scaffold(
+      body: ListView(
+        padding: const EdgeInsets.all(16),
+        children: [
+          _buildPetCategories(),
+          const SizedBox(height: 30),
+          _sectionHeader("Adopt pet", "See all"),
+          const SizedBox(height: 20),
+          _buildAdoptPetSection(),
+          const SizedBox(height: 30),
+          _sectionHeader("Suggestions", ""),
+          const SizedBox(height: 20),
+          _buildSuggestionsSection(suggestions, likedSuggestions, 'suggestion1'),
+          const SizedBox(height: 20),
+          _buildSuggestionsSection(suggestions2, likedSuggestions2, 'suggestion2'),
+        ],
+      ),
+    );
+  }
+
+  Widget _iconButton(IconData icon) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 8),
+      child: Container(
+        padding: const EdgeInsets.all(8),
+        decoration: BoxDecoration(
+          color: const Color(0xFFF5F5F5),
+          borderRadius: BorderRadius.circular(12),
+        ),
+        child: Icon(icon, color: Colors.black),
+      ),
+    );
+  }
+
+  Widget _buildPetCategories() {
     return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Container(
-            //-------------------------------------------animal lovers community header
-            margin: EdgeInsets.all(20),
-            padding: EdgeInsets.all(16),
-            decoration: BoxDecoration(
-                color: Theme.of(context).colorScheme.primary,
-                borderRadius: BorderRadius.circular(10)),
-            width: 455,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                ConstrainedBox(
-                  constraints: BoxConstraints(maxWidth: 200),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: const [
+            Text(
+              "Pet Categories",
+              style: TextStyle(fontWeight: FontWeight.bold, color: Colors.purple, fontSize: 22),
+            ),
+            Text(
+              "More Category",
+              style: TextStyle(color: Colors.grey, fontSize: 14),
+            ),
+          ],
+        ),
+        const SizedBox(height: 20),
+        Row(
+          children: categories.map((category) {
+            final isSelected = category == selectedCategory;
+            return Expanded(
+              child: GestureDetector(
+                onTap: () => setState(() => selectedCategory = category),
+                child: Container(
+                  padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 18),
+                  margin: const EdgeInsets.symmetric(horizontal: 8),
+                  decoration: BoxDecoration(
+                    color: isSelected ? Colors.purple : Colors.grey.shade200,
+                    borderRadius: BorderRadius.circular(25),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.grey.withValues(alpha: 0.4),
+                        spreadRadius: 2,
+                        blurRadius: 8,
+                        offset: const Offset(0, 4),
+                      ),
+                    ],
+                  ),
+                  alignment: Alignment.center,
                   child: Text(
-                    "Join our animal lovers community",
+                    category,
                     style: TextStyle(
-                        fontWeight: FontWeight.w600,
-                        fontSize: 20,
-                        color: Colors.white),
+                      color: isSelected ? Colors.white : Colors.black,
+                      fontWeight: FontWeight.w600,
+                      fontSize: 14,
+                    ),
                   ),
                 ),
-                SizedBox(height: 12),
-                ElevatedButton(
-                  onPressed: () {},
-                  style: ElevatedButton.styleFrom(
-                      shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(8))),
-                  child: Text("Join"),
-                )
-              ],
-            )),
-
-        Align(
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.only(left: 20.0),
-            child: Text(
-              "What are you looking for?",
-              style: TextStyle(
-                  fontWeight: FontWeight.w600,
-                  fontSize: 15,
-                  color: Theme.of(context).colorScheme.primary),
-            ),
-          ),
+              ),
+            );
+          }).toList(),
         ),
-        SizedBox(
-          height: 180,
-          child: ListView(
-            primary: false,
-            padding: EdgeInsets.only(top: 15.0),
-            scrollDirection: Axis.horizontal,
+      ],
+    );
+  }
+
+  Widget _sectionHeader(String title, String action) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Text(title, style: const TextStyle(fontWeight: FontWeight.bold, color: Colors.purple, fontSize: 22)),
+        Text(action, style: TextStyle(color: Colors.grey.shade600, fontSize: 16)),
+      ],
+    );
+  }
+
+  Widget _buildAdoptPetSection() {
+    return SingleChildScrollView(
+      scrollDirection: Axis.horizontal,
+      child: Row(
+        children: adoptPets.asMap().entries.map((entry) {
+          int index = entry.key;
+          var pet = entry.value;
+          return Padding(
+            padding: const EdgeInsets.only(right: 20),
+            child: PetCard(
+              name: pet['name']!,
+              imagePath: pet['image']!,
+              isLiked: likedPets[index],
+              onHeartTap: () => _toggleHeart(index, 'adopt'),
+              gender: pet['gender']!,
+            ),
+          );
+        }).toList(),
+      ),
+    );
+  }
+
+  Widget _buildSuggestionsSection(List<Map<String, String>> list, List<bool> likes, String section) {
+    return SingleChildScrollView(
+      scrollDirection: Axis.horizontal,
+      child: Row(
+        children: list.asMap().entries.map((entry) {
+          int index = entry.key;
+          var pet = entry.value;
+          return Padding(
+            padding: const EdgeInsets.only(right: 20),
+            child: PetCard(
+              name: pet['name']!,
+              imagePath: pet['image']!,
+              isLiked: likes[index],
+              onHeartTap: () => _toggleHeart(index, section),
+              gender: pet['gender']!,
+            ),
+          );
+        }).toList(),
+      ),
+    );
+  }
+}
+
+class PetCard extends StatelessWidget {
+  final String name;
+  final String imagePath;
+  final bool isLiked;
+  final VoidCallback onHeartTap;
+  final String gender;
+
+  const PetCard({
+    super.key,
+    required this.name,
+    required this.imagePath,
+    required this.isLiked,
+    required this.onHeartTap,
+    required this.gender,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: 220,
+      height: 250,
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.shade300,
+            blurRadius: 6,
+            spreadRadius: 2,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Stack(
             children: [
-              Column(
-                children: [
-                  Container(
-                    margin: EdgeInsets.symmetric(horizontal: 10.0),
-                    padding: EdgeInsets.all(10),
-                    decoration: BoxDecoration(
-                        color: const Color(0xFFA0A2DA),
-                        borderRadius: BorderRadius.circular(40),
-                        border: Border.all(width: 1)),
-                    width: 110,
-                    height: 107,
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(25),
-                      child: Image(
-                          fit: BoxFit.cover,
-                          image: AssetImage('assets/images/store.png')),
-                    ),
-                  ),
-                  Container(
-                    alignment: Alignment.center,
-                    padding: EdgeInsets.only(top: 5),
-                    child: Text(
-                      "Pet Store",
-                      style: TextStyle(
-                        fontSize: 15,
-                      ),
-                    ),
-                  ),
-                ],
+              ClipRRect(
+                borderRadius: const BorderRadius.vertical(top: Radius.circular(20)),
+                child: Image.asset(imagePath, width: 220, height: 200, fit: BoxFit.cover),
               ),
-              Column(
-                children: [
-                  Container(
-                    margin: EdgeInsets.symmetric(horizontal: 10.0),
-                    padding: EdgeInsets.all(10),
+              Positioned(
+                right: 12,
+                top: 12,
+                child: GestureDetector(
+                  onTap: onHeartTap,
+                  child: Container(
+                    padding: const EdgeInsets.all(4),
                     decoration: BoxDecoration(
-                      color: const Color(0xFFF18487),
-                      borderRadius: BorderRadius.circular(40),
-                      border: Border.all(width: 1),
+                      color: const Color(0xFFF5F5F5),
+                      borderRadius: BorderRadius.circular(12),
                     ),
-                    width: 110,
-                    height: 107,
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(40),
-                      child: Image(
-                          fit: BoxFit.cover,
-                          image: AssetImage('assets/images/grooming.png')),
+                    child: Icon(
+                      isLiked ? Icons.favorite : Icons.favorite_border,
+                      color: Colors.red,
+                      size: 20,
                     ),
                   ),
-                  Container(
-                    alignment: Alignment.center,
-                    padding: EdgeInsets.only(top: 5),
-                    child: Text(
-                      "Pet Grooming",
-                      style: TextStyle(
-                        fontSize: 15,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              Column(
-                children: [
-                  Container(
-                    margin: EdgeInsets.symmetric(horizontal: 10.0),
-                    padding: EdgeInsets.all(10),
-                    decoration: BoxDecoration(
-                        color: Color(0xFFEF9E1B),
-                        borderRadius: BorderRadius.circular(40),
-                        border: Border.all(width: 1)),
-                    width: 110,
-                    height: 107,
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(40),
-                      child: Image(
-                          fit: BoxFit.cover,
-                          image: AssetImage('assets/images/boarding.png')),
-                    ),
-                  ),
-                  Container(
-                    alignment: Alignment.center,
-                    padding: EdgeInsets.only(top: 5),
-                    child: Text(
-                      "Pet Boarding",
-                      style: TextStyle(
-                        fontSize: 15,
-                      ),
-                    ),
-                  ),
-                ],
+                ),
               ),
             ],
           ),
-        ),
-      ],
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  name,
+                  style: const TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 18,
+                  ),
+                ),
+                Container(
+                  margin: const EdgeInsets.only(top: 4),
+                  padding: const EdgeInsets.all(6),
+                  decoration: BoxDecoration(
+                    color: const Color(0xFFF5F5F5),
+                    shape: BoxShape.circle,
+                  ),
+                  child: Icon(
+                    gender == 'Female' ? Icons.female : Icons.male,
+                    size: 20,
+                    color: Colors.purple,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
     );
   }
 }
